# architecture/api-contract-openapi.yaml
openapi: 3.0.3
info:
  title: iReg Moto BF API
  description: >
    API réglementaire pour la conformité des deux-roues motorisés au Burkina Faso.
    Plateforme SaaS interfaçant importateurs, distributeurs, unités d'assemblage,
    détaillants et le ministère des transports (DRCTT).
    Supporte REST (mutations métier) et GraphQL (analytics).
  version: 1.0.0
  contact:
    name: iReg Moto BF Technical Team
    email: tech@ireg-moto.bf
  license:
    name: Proprietary
    url: https://ireg-moto.bf/licence
servers:
  - url: https://api.ireg-moto.bf/v1
    description: Production
  - url: https://api-staging.ireg-moto.bf/v1
    description: Staging
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /sales:
    get:
      tags: [Ventes]
      summary: Liste des ventes
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, PENDING_PAYMENT, PAID, PARTIALLY_PAID, CANCELLED, REFUNDED]
        - name: actorId
          in: query
          schema: { type: string, format: uuid }
        - name: clientId
          in: query
          schema: { type: string, format: uuid }
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
        - name: paymentMethod
          in: query
          schema:
            type: string
            enum: [CASH, BANK_TRANSFER, MOBILE_MONEY, CHEQUE, CREDIT, INSTALLMENT]
        - name: minAmount
          in: query
          schema: { type: number }
        - name: maxAmount
          in: query
          schema: { type: number }
      responses:
        '200':
          description: Liste des ventes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSales'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Ventes]
      summary: Enregistrer une vente
      description: >
        Enregistre une vente de véhicule avec facturation OHADA.
        Vérifie automatiquement : blacklist, conformité, stock disponible.
        Déclenche la détection de fraude post-vente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSaleRequest'
      responses:
        '201':
          description: Vente enregistrée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: Vente bloquée (blacklist, non-conformité, stock indisponible)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /sales/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }

    get:
      tags: [Ventes]
      summary: Détail d'une vente
      responses:
        '200':
          description: Détail vente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Ventes]
      summary: Modifier une vente (avant finalisation)
      description: Modification possible uniquement si statut DRAFT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSaleRequest'
      responses:
        '200':
          description: Vente mise à jour
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Vente non modifiable (statut avancé)
  /sales/{id}/cancel:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Ventes]
      summary: Annuler une vente
      description: >
        Annule une vente et restaure le stock.
        Nécessite une justification.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 10
      responses:
        '200':
          description: Vente annulée
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Vente déjà annulée ou trop ancienne (> 30 jours)
  /sales/{id}/invoice:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Ventes]
      summary: Télécharger la facture OHADA
      description: >
        Retourne la facture PDF signée électroniquement
        conforme aux normes OHADA.
      responses:
        '200':
          description: Facture PDF
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==========================================================================
  # PRICING (MODULE D)
  # ==========================================================================
  /pricing/history:
    get:
      tags: [Pricing]
      summary: Historique des prix
      description: >
        Retourne l'historique des prix d'achat et de vente
        avec calcul automatique des marges.
      parameters:
        - name: actorId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: vehicleCategoryId
          in: query
          schema: { type: string, format: uuid }
        - name: manufacturer
          in: query
          schema: { type: string }
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Historique des prix
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PricingHistoryEntry'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Pricing]
      summary: Enregistrer un prix d'achat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePurchasePriceRequest'
      responses:
        '201':
          description: Prix enregistré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchasePrice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /security/blacklist:
    get:
      tags: [Securite]
      summary: Liste noire de sécurité
      description: Entités blacklistées (véhicules, clients, acteurs).
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [VEHICLE, CLIENT, ACTOR, USER, VIN_PATTERN]
        - name: search
          in: query
          schema: { type: string }
        - name: isActive
          in: query
          schema: { type: boolean, default: true }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Liste noire
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SecurityBlacklistEntry'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Securite]
      summary: Ajouter une entrée à la liste noire
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSecurityBlacklistRequest'
      responses:
        '201':
          description: Entrée ajoutée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityBlacklistEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT d'accès (RS256)
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clé API pour les intégrations système (ministère, DGI)

  # ========================================================================
  # RESPONSES
  # ========================================================================
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Non authentifié — Token invalide ou expiré
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Accès interdit — Permissions insuffisantes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required: [statusCode, message]
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: "Le VIN fourni n'est pas valide"
        code:
          type: string
          example: INVALID_VIN
        details:
          type: object
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: /api/v1/vehicles
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean

    # -------------------------------------------------------------------------
    # Auth
    # -------------------------------------------------------------------------
...
    Sale:
      type: object
      properties:
        id: { type: string }
        actorId: { type: string }
        clientId: { type: string }
        clientName: { type: string }
        status:
          type: string
          enum: [DRAFT, PENDING_PAYMENT, PAID, PARTIALLY_PAID, CANCELLED, REFUNDED]
        paymentMethod:
          type: string
          enum: [CASH, BANK_TRANSFER, MOBILE_MONEY, CHEQUE, CREDIT, INSTALLMENT]
        subtotalAmount: { type: number }
        discountAmount: { type: number }
        taxAmount: { type: number }
        totalAmount: { type: number }
        currencyCode: { type: string }
        saleDate:
          type: string
          format: date-time
        invoiceNumber: { type: string }
        itemCount: { type: integer }
        createdAt:
          type: string
          format: date-time
    SaleDetail:
      allOf:
        - $ref: '#/components/schemas/Sale'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/SaleItem'
            taxBreakdown:
              type: object
            invoice:
              type: object
              properties:
                id: { type: string }
                invoiceNumber: { type: string }
                status: { type: string }
                pdfUrl: { type: string }
    PaginatedSales:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Sale'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CreateSaleRequest:
      type: object
      required: [clientId, paymentMethod, items]
      properties:
        clientId:
          type: string
          format: uuid
        warehouseId:
          type: string
          format: uuid
        paymentMethod:
          type: string
          enum: [CASH, BANK_TRANSFER, MOBILE_MONEY, CHEQUE, CREDIT, INSTALLMENT]
        paymentReference: { type: string }
        discountAmount: { type: number }
        notes: { type: string }
        items:
          type: array
          items:
            type: object
            required: [vehicleId, unitPrice]
            properties:
              vehicleId:
                type: string
                format: uuid
              unitPrice: { type: number }
              quantity:
                type: integer
                default: 1
              discountAmount: { type: number }
        saleLocation:
          type: object
          properties:
            lat: { type: number }
            lng: { type: number }
    UpdateSaleRequest:
      type: object
      properties:
        discountAmount: { type: number }
        notes: { type: string }
    SaleItem:
      type: object
      properties:
        id: { type: string }
        vehicleId: { type: string }
        vehicleVin: { type: string }
        vehicleModel: { type: string }
        unitPrice: { type: number }
        quantity: { type: integer }
        lineTotal: { type: number }
        taxAmount: { type: number }
        marginAmount: { type: number }
        marginRate: { type: number }

    # -------------------------------------------------------------------------
    # Pricing
    # -------------------------------------------------------------------------
    PurchasePrice:
      type: object
      properties:
        id: { type: string }
        actorId: { type: string }
        vehicleCategoryId: { type: string }
        manufacturer: { type: string }
        model: { type: string }
        modelYear: { type: integer }
        unitPriceForeign: { type: number }
        foreignCurrencyCode: { type: string }
        exchangeRate: { type: number }
        unitPriceXof: { type: number }
        freightCost: { type: number }
        insuranceCost: { type: number }
        customsDuties: { type: number }
        otherCosts: { type: number }
        totalLandedCost: { type: number }
        supplierName: { type: string }
        supplierCountry: { type: string }
        effectiveDate:
          type: string
          format: date
    CreatePurchasePriceRequest:
      type: object
      required: [actorId, unitPriceForeign, foreignCurrencyCode, exchangeRate, totalLandedCost]
      properties:
        actorId:
          type: string
          format: uuid
        vehicleCategoryId:
          type: string
          format: uuid
        manufacturer: { type: string }
        model: { type: string }
        modelYear: { type: integer }
        unitPriceForeign: { type: number }
        foreignCurrencyCode:
          type: string
          default: EUR
        exchangeRate: { type: number }
        freightCost: { type: number }
        insuranceCost: { type: number }
        customsDuties: { type: number }
        otherCosts: { type: number }
        totalLandedCost: { type: number }
        supplierName: { type: string }
        supplierCountry: { type: string }
        effectiveDate:
          type: string
          format: date
    PricingHistoryEntry:
      type: object
      properties:
        id: { type: string }
        actorId: { type: string }
        vehicleCategory: { type: string }
        manufacturer: { type: string }
        model: { type: string }
        purchasePrice: { type: number }
        sellingPrice: { type: number }
        marginAmount: { type: number }
        marginRate: { type: number }
        isAnomaly: { type: boolean }
        anomalyType: { type: string }
        effectiveDate:
          type: string
          format: date

    # -------------------------------------------------------------------------
    # Conformité
    # -------------------------------------------------------------------------
...
    SecurityBlacklistEntry:
      type: object
      properties:
        id: { type: string }
        blacklistType: { type: string }
        entityId: { type: string }
        entityIdentifier: { type: string }
        reason: { type: string }
        reasonCategory: { type: string }
        source: { type: string }
        blacklistedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        isActive: { type: boolean }
    CreateSecurityBlacklistRequest:
      type: object
      required: [blacklistType, entityIdentifier, reason, reasonCategory]
      properties:
        blacklistType:
          type: string
          enum: [VEHICLE, CLIENT, ACTOR, USER, VIN_PATTERN]
        entityId:
          type: string
          format: uuid
        entityIdentifier: { type: string }
        reason: { type: string }
        reasonCategory:
          type: string
          enum: [FRAUD, THEFT, COUNTERFEIT, SANCTION]
        sourceReference: { type: string }
        expiresAt:
          type: string
          format: date-time
...